+++ demo/src/lineDemo.ts
@@ -6,22 +6,20 @@ import boundaryBox from './boundaryBox'
 const boundary = 10
 
 function lineDemo(ci: CanvasInfo) {
-  const { ctx, canvas } = ci
-
   boundaryBox(ci, boundary)
 
   // draw lines
-  line(ctx, {
+  line(ci.ctx, {
     x1: boundary,
     y1: boundary,
-    x2: canvas.width - boundary,
-    y2: canvas.height - boundary,
+    x2: ci.width - boundary,
+    y2: ci.height - boundary,
     color: 'red',
   })
-  line(ctx, {
+  line(ci.ctx, {
     x1: boundary,
-    y1: canvas.height - boundary,
-    x2: canvas.width - boundary,
+    y1: ci.height - boundary,
+    x2: ci.width - boundary,
     y2: boundary,
     color: 'green',
     dashed: true,